requirements.txt
Flask
mysql-connector-python==8.0.20